proj: Initial project setup with basic file stubs.
PConf is a  perl based script to configure Linux kernel modules
which are supposed to be mainly build out of tree (that means;
no menuconf). PConf uses various input scripts where it bases
its output on. See http://michelmegens.net/pconf/ for the full
documentation.

Last but not least, it can be used to export a module into the
Linux tree, to configure it using menuconf. This could be used
if you want to build it into the actual kernel image.

Signed-off-by: Michel Megens <dev@michelmegens.net>