pconf: args: added some serious argument parsing.
The previous argument parsing was just amateuristic tinkering. Now,
on the other hand, we are using a CPAN module called Getopts::Mixed.
This module (which is not installed by default, ofcourse) parses all
arguments smoothly and gives you back the value assigned to it, if
there was a value at all.

Signed-off-by: Michel Megens <dev@michelmegens.net>