template: added template directory and file stubs
File templates will be used to build up a Kbuild file (if building
out-of-tree) or a Makefile (if building in-tree). The template files
use a 'search and replace' policy. Certain parts of the file will
marked with a @unique_string@ marker. These parts will be replaced
by the script generator to build the actual file.

Signed-off-by: Michel Megens <dev@michelmegens.net>